Setup compatibility with composerlib
 * Imports composerlib as an external library to create musical composition tools. From now on, composer is a tool that integrates MIDI support into composerlib.

* Integrates the `melody_problem` class with midi_io to create a problem from a midi file.
