+++ examples/my_optimizer.cpp
@@ -2,9 +2,9 @@
 // Created by Carla on 20/06/2022.
 //
 
-#include "composer/optimizer.hpp"
-#include "composer/genetic_algorithm.h"
-#include "composer/nsga_ii.h"
+#include "composerlib/optimizer.hpp"
+#include "composerlib/genetic_algorithm.h"
+#include "composerlib/nsga_ii.h"
 
 using namespace composer;
 
@@ -16,7 +16,7 @@ void solver() {
     double mutation_strength = 0.3;
     int population_size = 1000;
     size_t max_iterations = 30;
-    size_t num_executions = 100;
+    size_t num_executions = 10;
 
     std::pair<double, double> target = {-1., 0.};
     const melody_problem p(melody_problem::problem_type::random, 100, target);
