+++ source/CMakeLists.txt
@@ -4,11 +4,6 @@
 # Threads
 find_package(Threads)
 
-# Ranges library
-CPMAddPackage(NAME range-v3 URL https://github.com/ericniebler/range-v3/archive/0.11.0.zip VERSION 0.11.0 DOWNLOAD_ONLY TRUE)
-add_library(range-v3 INTERFACE IMPORTED)
-target_include_directories(range-v3 INTERFACE "${range-v3_SOURCE_DIR}/include")
-
 # Infra library
 FetchContent_Declare(
         infra
@@ -51,50 +46,40 @@ if(NOT midifile_POPULATED)
 endif()
 
 target_include_directories(midifile PUBLIC "${midifile_SOURCE_DIR}/include")
+
+# Composerlib library
 FetchContent_Declare(
-        pareto
-        GIT_REPOSITORY https://github.com/alandefreitas/pareto.git
-        GIT_TAG c133aa6c3be682d12e6fcbfa26a80da28ebb8b70
+        composerlib
+        GIT_REPOSITORY https://github.com/composernew/composerlib.git
+        GIT_TAG 798ad10b67b3056074d882aebd20f1e424ac537b
 )
 
-FetchContent_GetProperties(pareto)
-if (NOT pareto_POPULATED)
-    FetchContent_Populate(pareto)
-    add_subdirectory(${pareto_SOURCE_DIR} ${pareto_BINARY_DIR} EXCLUDE_FROM_ALL)
-    set(pareto_BUILD_PARETO_WITH_PMR_BY_DEFAULT OFF)
+FetchContent_GetProperties(composerlib)
+if (NOT composerlib_POPULATED)
+    FetchContent_Populate(composerlib)
+    add_subdirectory(${composerlib_SOURCE_DIR} ${composerlib_BINARY_DIR} EXCLUDE_FROM_ALL)
+    set(composerlib_BUILD_PARETO_WITH_PMR_BY_DEFAULT OFF)
 endif ()
 
 #######################################################
 ### Library                                         ###
 #######################################################
-add_library(composerlib
-        composer/harmony.cpp
-        composer/harmony.h
-        composer/melody.cpp
-        composer/melody.h
-        composer/objective_parameters.cpp
-        composer/objective_parameters.h
-        composer/synth.h
-        composer/synth.cpp
-        composer/midi_stream.cpp
-        composer/midi_stream.h
-        composer/midi_processor.cpp
-        composer/midi_processor.h
-        composer/melody_problem.cpp
-        composer/melody_problem.h
-        composer/genetic_algorithm.cpp
-        composer/genetic_algorithm.h
-        composer/nsga_ii.cpp
-        composer/nsga_ii.h
-        composer/optimizer.hpp
-        composer/optimizer.h
-        composer/midi_io.cpp
-        composer/midi_io.h
+add_library(composermidi
+        composermidi/harmony.cpp
+        composermidi/harmony.h
+        composermidi/synth.h
+        composermidi/synth.cpp
+        composermidi/midi_stream.cpp
+        composermidi/midi_stream.h
+        composermidi/midi_processor.cpp
+        composermidi/midi_processor.h
+        composermidi/midi_io.cpp
+        composermidi/midi_io.h
 )
 
-target_include_directories(composerlib
+target_include_directories(composermidi
         PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>
                $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>)
-target_link_libraries(composerlib PUBLIC Threads::Threads range-v3 libq libqio cycfi::infra midifile pareto)
-target_pedantic_options(composerlib)
-target_exception_options(composerlib)
+target_link_libraries(composermidi PUBLIC Threads::Threads libq libqio cycfi::infra midifile composerlib)
+target_pedantic_options(composermidi)
+target_exception_options(composermidi)
