--- source/composer/objective_parameters.cpp
@@ -1,5 +1,0 @@
-//
-// Created by Alan Freitas on 12/14/20.
-//
-
-#include "objective_parameters.h"
