--- source/composer/objective_parameters.h
@@ -1,12 +1,0 @@
-//
-// Created by Alan Freitas on 12/14/20.
-//
-
-#ifndef COMPOSER_OBJECTIVE_PARAMETERS_H
-#define COMPOSER_OBJECTIVE_PARAMETERS_H
-
-namespace composer {
-    class objective_parameters {};
-}
-
-#endif // COMPOSER_OBJECTIVE_PARAMETERS_H
