+++ source/composermidi/harmony.cpp
