+++ source/composermidi/midi_stream.cpp
