+++ tests/unit_tests/CMakeLists.txt
@@ -26,7 +26,7 @@ include(${Catch2_SOURCE_DIR}/contrib/Catch.cmake)
 #######################################################
 
 add_executable(ut_midi_stream ut_midi_stream.cpp)
-target_link_libraries(ut_midi_stream PUBLIC composerlib Catch2)
+target_link_libraries(ut_midi_stream PUBLIC composermidi Catch2)
 
 if (MSVC)
     set_target_properties(ut_midi_stream PROPERTIES
